Created board controller class, refactored code to use it correctly
